+++ openfga_sdk/sync/oauth2.py
@@ -80,7 +80,7 @@ class OAuth2Client:
             {
                 "Accept": "application/json",
                 "Content-Type": "application/x-www-form-urlencoded",
-                "User-Agent": "openfga-sdk (python) 0.4.3",
+                "User-Agent": "openfga-sdk (python) 0.5.0",
             }
         )