+++ test/test_client_sync.py
@@ -39,6 +39,7 @@ from openfga_sdk.models.check_response import CheckResponse
 from openfga_sdk.models.create_store_request import CreateStoreRequest
 from openfga_sdk.models.create_store_response import CreateStoreResponse
 from openfga_sdk.models.expand_response import ExpandResponse
+from openfga_sdk.models.fga_object import FgaObject
 from openfga_sdk.models.get_store_response import GetStoreResponse
 from openfga_sdk.models.leaf import Leaf
 from openfga_sdk.models.list_objects_response import ListObjectsResponse
@@ -974,7 +975,6 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
             mock_response("{}", 200),
             mock_response("{}", 200),
             mock_response("{}", 200),
-            mock_response("{}", 200),
         ]
         configuration = self.configuration
         configuration.store_id = store_id
@@ -1043,15 +1043,7 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
                     ),
                 ],
             )
-            self.assertEqual(mock_request.call_count, 4)
-            mock_request.assert_any_call(
-                "GET",
-                "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/authorization-models/01G5JAVJ41T49E9TT3SKVS7X1J",
-                headers=ANY,
-                query_params=[],
-                _preload_content=ANY,
-                _request_timeout=None,
-            )
+            self.assertEqual(mock_request.call_count, 3)
             mock_request.assert_any_call(
                 "POST",
                 "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/write",
@@ -1126,7 +1118,6 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
             mock_response("{}", 200),
             mock_response("{}", 200),
             mock_response("{}", 200),
-            mock_response("{}", 200),
         ]
         configuration = self.configuration
         configuration.store_id = store_id
@@ -1195,15 +1186,7 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
                     ),
                 ],
             )
-            self.assertEqual(mock_request.call_count, 4)
-            mock_request.assert_any_call(
-                "GET",
-                "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/authorization-models/01G5JAVJ41T49E9TT3SKVS7X1J",
-                headers=ANY,
-                query_params=[],
-                _preload_content=ANY,
-                _request_timeout=None,
-            )
+            self.assertEqual(mock_request.call_count, 3)
             mock_request.assert_any_call(
                 "POST",
                 "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/write",
@@ -1346,15 +1329,7 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
                     ),
                 ],
             )
-            self.assertEqual(mock_request.call_count, 3)
-            mock_request.assert_any_call(
-                "GET",
-                "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/authorization-models/01G5JAVJ41T49E9TT3SKVS7X1J",
-                headers=ANY,
-                query_params=[],
-                _preload_content=ANY,
-                _request_timeout=None,
-            )
+            self.assertEqual(mock_request.call_count, 2)
             mock_request.assert_any_call(
                 "POST",
                 "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/write",
@@ -1417,7 +1392,6 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
         """
 
         mock_request.side_effect = [
-            mock_response("{}", 200),
             mock_response("{}", 200),
             ValidationException(http_resp=http_mock_response(response_body, 400)),
             mock_response("{}", 200),
@@ -1496,15 +1470,7 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
                     error=None,
                 ),
             )
-            self.assertEqual(mock_request.call_count, 4)
-            mock_request.assert_any_call(
-                "GET",
-                "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/authorization-models/01G5JAVJ41T49E9TT3SKVS7X1J",
-                headers=ANY,
-                query_params=[],
-                _preload_content=ANY,
-                _request_timeout=None,
-            )
+            self.assertEqual(mock_request.call_count, 3)
             mock_request.assert_any_call(
                 "POST",
                 "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/write",
@@ -1577,7 +1543,6 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
         """
         mock_request.side_effect = [
             mock_response("{}", 200),
-            mock_response("{}", 200),
         ]
         configuration = self.configuration
         configuration.store_id = store_id
@@ -1602,14 +1567,6 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
                     "authorization_model_id": "01G5JAVJ41T49E9TT3SKVS7X1J",
                     "transaction": transaction,
                 },
-            )
-            mock_request.assert_any_call(
-                "GET",
-                "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/authorization-models/01G5JAVJ41T49E9TT3SKVS7X1J",
-                headers=ANY,
-                query_params=[],
-                _preload_content=ANY,
-                _request_timeout=None,
             )
             mock_request.assert_any_call(
                 "POST",
@@ -1771,7 +1728,7 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
         configuration = self.configuration
         configuration.store_id = store_id
         with OpenFgaClient(configuration) as api_client:
-            with self.assertRaises(UnauthorizedException) as api_exception:
+            with self.assertRaises(UnauthorizedException):
                 body = ClientWriteRequest(
                     writes=[
                         ClientTuple(
@@ -1792,19 +1749,31 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
                     },
                 )
 
-            self.assertIsInstance(api_exception.exception, UnauthorizedException)
-            mock_request.assert_called()
-            self.assertEqual(mock_request.call_count, 1)
+                mock_request.assert_called()
+                self.assertEqual(mock_request.call_count, 1)
 
-            mock_request.assert_called_once_with(
-                "GET",
-                "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/authorization-models/01G5JAVJ41T49E9TT3SKVS7X1J",
-                headers=ANY,
-                query_params=[],
-                _preload_content=ANY,
-                _request_timeout=None,
-            )
-            api_client.close()
+                mock_request.assert_called_once_with(
+                    "POST",
+                    "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/write",
+                    headers=ANY,
+                    query_params=[],
+                    post_params=[],
+                    body={
+                        "writes": {
+                            "tuple_keys": [
+                                {
+                                    "user": "user:81684243-9356-4421-8fbf-a4f8d36aa31b",
+                                    "relation": "reader",
+                                    "object": "document:2021-budget",
+                                }
+                            ]
+                        },
+                        "authorization_model_id": "01G5JAVJ41T49E9TT3SKVS7X1J",
+                    },
+                    _preload_content=ANY,
+                    _request_timeout=ANY,
+                )
+                api_client.close()
 
     @patch.object(rest.RESTClientObject, "request")
     def test_check(self, mock_request):
@@ -1918,7 +1887,6 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
         # First, mock the response
         response_body = '{"allowed": true, "resolution": "1234"}'
         mock_request.side_effect = [
-            mock_response("{}", 200),
             mock_response(response_body, 200),
         ]
         body = ClientCheckRequest(
@@ -1939,14 +1907,6 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
             self.assertTrue(api_response[0].allowed)
             self.assertEqual(api_response[0].request, body)
             # Make sure the API was called with the right data
-            mock_request.assert_any_call(
-                "GET",
-                "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/authorization-models/01GXSA8YR785C4FYS3C0RTG7B1",
-                headers=ANY,
-                query_params=[],
-                _preload_content=ANY,
-                _request_timeout=None,
-            )
             mock_request.assert_any_call(
                 "POST",
                 "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/check",
@@ -1975,7 +1935,6 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
 
         # First, mock the response
         mock_request.side_effect = [
-            mock_response("{}", 200),
             mock_response('{"allowed": true, "resolution": "1234"}', 200),
             mock_response('{"allowed": false, "resolution": "1234"}', 200),
             mock_response('{"allowed": true, "resolution": "1234"}', 200),
@@ -2017,14 +1976,6 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
             self.assertTrue(api_response[2].allowed)
             self.assertEqual(api_response[2].request, body3)
             # Make sure the API was called with the right data
-            mock_request.assert_any_call(
-                "GET",
-                "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/authorization-models/01GXSA8YR785C4FYS3C0RTG7B1",
-                headers=ANY,
-                query_params=[],
-                _preload_content=ANY,
-                _request_timeout=None,
-            )
             mock_request.assert_any_call(
                 "POST",
                 "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/check",
@@ -2093,7 +2044,6 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
 
         # First, mock the response
         mock_request.side_effect = [
-            mock_response("{}", 200),
             mock_response('{"allowed": true, "resolution": "1234"}', 200),
             ValidationException(http_resp=http_mock_response(response_body, 400)),
             mock_response('{"allowed": false, "resolution": "1234"}', 200),
@@ -2138,14 +2088,6 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
             self.assertFalse(api_response[2].allowed)
             self.assertEqual(api_response[2].request, body3)
             # Make sure the API was called with the right data
-            mock_request.assert_any_call(
-                "GET",
-                "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/authorization-models/01GXSA8YR785C4FYS3C0RTG7B1",
-                headers=ANY,
-                query_params=[],
-                _preload_content=ANY,
-                _request_timeout=None,
-            )
             mock_request.assert_any_call(
                 "POST",
                 "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/check",
@@ -2357,7 +2299,6 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
 
         # First, mock the response
         mock_request.side_effect = [
-            mock_response("{}", 200),
             mock_response('{"allowed": true, "resolution": "1234"}', 200),
             mock_response('{"allowed": false, "resolution": "1234"}', 200),
             mock_response('{"allowed": true, "resolution": "1234"}', 200),
@@ -2376,14 +2317,6 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
             self.assertEqual(api_response, ["reader", "viewer"])
 
             # Make sure the API was called with the right data
-            mock_request.assert_any_call(
-                "GET",
-                "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/authorization-models/01GXSA8YR785C4FYS3C0RTG7B1",
-                headers=ANY,
-                query_params=[],
-                _preload_content=ANY,
-                _request_timeout=None,
-            )
             mock_request.assert_any_call(
                 "POST",
                 "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/check",
@@ -2459,16 +2392,7 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
 
             self.assertIsInstance(api_exception.exception, UnauthorizedException)
             mock_request.assert_called()
-            self.assertEqual(mock_request.call_count, 1)
-
-            mock_request.assert_called_once_with(
-                "GET",
-                "http://api.fga.example/stores/01YCP46JKYM8FJCQ37NMBYHE5X/authorization-models/01GXSA8YR785C4FYS3C0RTG7B1",
-                headers=ANY,
-                query_params=[],
-                _preload_content=ANY,
-                _request_timeout=None,
-            )
+            self.assertEqual(mock_request.call_count, 3)
             api_client.close()
 
     @patch.object(rest.RESTClientObject, "request")
@@ -2478,7 +2402,6 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
         """
 
         response_body = """{
-  "excluded_users": [],
   "users": [
     {
       "object": {
@@ -2508,7 +2431,7 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
 
         with OpenFgaClient(configuration) as api_client:
             body = ClientListUsersRequest()
-            body.object = "document:2021-budget"
+            body.object = FgaObject(type="document", id="2021-budget")
             body.relation = "can_read"
             body.user_filters = [
                 UserTypeFilter(type="user"),
@@ -2564,7 +2487,7 @@ class TestOpenFgaClient(IsolatedAsyncioTestCase):
                 post_params=[],
                 body={
                     "authorization_model_id": "01G5JAVJ41T49E9TT3SKVS7X1J",
-                    "object": "document:2021-budget",
+                    "object": {"id": "2021-budget", "type": "document"},
                     "relation": "can_read",
                     "user_filters": [
                         {"type": "user"},